requirements.txt
flask
gtts
openai
requests
google-api-python-client
elevenlabs
python-dotenv
